fix[tree]: adjust generalized logic
